requirements.txt
# Delta-Neutral Controller — зависимости
# Primary shareable path: dry-run + unit tests + DN scripts.
# Live integrations are optional and require exchange credentials.

pyyaml>=6.0
python-dotenv>=1.0
aiohttp>=3.10.0
curl_cffi>=0.11.0
pydantic>=2.9.0

# Extended SDK (x10) runtime — для extended_adapter
fast-stark-crypto==0.3.8
sortedcontainers>=2.4.0
strenum>=0.4.15
tenacity>=9.1.2

# Variational runtime — для variational_adapter
eth-account>=0.12.0

# Optional Nado integration (legacy/non-default)
websockets>=12.0,<14.0
web3>=7.0.0
